backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.3

# Audio processing
pyaudio==0.2.13
sounddevice==0.4.6
webrtcvad==2.0.10

# Transcription & NLP
openai==1.10.0
transformers==4.37.2
numpy==1.26.3
torch==2.2.0

# Platform integration
zoom-sdk-python==0.1.7  # Note: This is a placeholder, actual package may differ
google-api-python-client==2.110.0
